requirements.txt
dj-database-url==0.4.2
Django==1.10.5
django-cors-headers==2.0.2
djangorestframework==3.5.3
psycopg2==2.6.2
python-dateutil==2.6.0
requests==2.12.5
six==1.10.0
